Refactor: useLike, LikeIcon 리팩토링
