+++ src/modules/CommentList/CommentCommon/index.ts
@@ -1,6 +1,7 @@
 export * from './Content.client';
 export * from './Header.client';
 export * from './LikeCount.client';
+export * from './LikeIcon.client';
 export * from './ReplyHideButton.client';
 export * from './ReplyShowButton.client';
 export * from './ReplySubmitButton.client';
